Bug 2004791 - Stub out the "state" getter on the backup service singleton, not its prototype. r=fchasen
I'm not sure why this helps, but it does---it seems that the UI otherwise
gets the 'real' backup state, even though it's stubbed out. (I'm wondering
if the BackupService prototype is different at some important point...?)

This change makes the restore-from-backup test fail if
backupFileCoarseLocation is not present, which is what I was confused
about in

  https://phabricator.services.mozilla.com/D274841#inline-1497362

It also resolves a failure in
test_turn_off_scheduled_backups_disables_encryption that would otherwise be
introduced in my patch for bug 1996249.

Differential Revision: https://phabricator.services.mozilla.com/D275517

UltraBlame original commit: 289cdbc2f8469a65f228db483b651580b8767be4