Bug 1809303 - Add "Copy link text" context menu item r=geckoview-reviewers,android-reviewers,android-l10n-reviewers,geckoview-api-reviewers,flod,ohall,tcampbell
This patch introduces a "Copy link text" option to the context menu. It adds the necessary `innerText` property from Gecko to the Android components layer, allowing the context menu feature to access and copy the visible text of a link to the clipboard.

Differential Revision: https://phabricator.services.mozilla.com/D275482

UltraBlame original commit: 4ff69f0b1422c75385cb893869cc256ca9405454