Bug 2002796 - Prefetch computation of gecko_taskgraph.files_changed.get_locally_changed_files r=jmaher,taskgraph-reviewers,jcristau
When get_locally_changed_files invokes git, it calls a subcommand that
takes ~700ms. This is quite slow and blocks the startup of several
taskgraph-related tasks.

We usually ask for the files changed in the main repo, so it maybe worth
prefetching that value in a separate thread. GIL is not an issue here as
we invoke a subprocess.

Differential Revision: https://phabricator.services.mozilla.com/D274297

UltraBlame original commit: 3e7a0e55ff57e7ae7fdb50f69cac7c6928d76bc7