Bug 2006348 - Refactor DefaultTabManagerControllerTest to use StandardTestDispatcher r=android-reviewers,giorga
This patch replaces the usage of `MainCoroutineRule` and `runTestOnMain` with `StandardTestDispatcher` and `runTest`.
To support this, `TabManagerController` is updated to accept an injected `mainDispatcher`, allowing tests to control the execution context. Use of explicit dispatchers in `TabManagementFragment` is removed in favor of the default values.

Differential Revision: https://phabricator.services.mozilla.com/D276674

UltraBlame original commit: 5aba73b79f3e14095da30753bc32e4f7c92ac0b8