Bug 2006032 - Expect that `/selection/shadow-dom/cross-shadow-boundary-1.html` passes if using draw snapshot r=TYLin
If the task uses the draw snapshot, accessible carets are not rendered
in the result and the test fails with the odd result of the accessible
caret.  So, we can expect the fail is only when the task does not use
the draw snapshot.

Differential Revision: https://phabricator.services.mozilla.com/D276586

UltraBlame original commit: efec677c3071d97e5ef0676833498fbc795bd842