Bug 1262425 - [devtools] Fix displaying shorthand computed properties whose value depends on longhand using CSS variable. r=devtools-reviewers,bomsy
When setting a longhand property value using CSS variable, the property value
we get for the shorthand from the rule CSSStyleDeclaration is an empty string.
In such case, we don't want to filter out the property if the "Browser styles"
checkbox is not checked.
In order to fix this, we check for the presence of the property in the CSSStyleDeclaration
using the chrome-only `hasLonghandProperty` method that is added in the previous commit
of this stack.
We also change a similar check to make sure we include the matched selector, even
if we can't display a meaningful value to the user (see Bug 2003264).

Because of these changes, browser_computed_custom_properties.js needed to be
updated. We change the `border` we were using to `outline` because it gives
us less longhand properties to check.

Differential Revision: https://phabricator.services.mozilla.com/D274586

UltraBlame original commit: 8ddcc220e631c9f0604578906aa9a745a2bb6db9