Bug 1999828: Part 1 - Split TryFoldingStubs into two passes. One to detect the shape (and later slot offset) and another to gather the values. r=iain
Preparation to being able to fold GuardShape followed by {Load|Store}{Fixed|Dynamic}Slot.
For this we split the detection of stubs with different shapes (and in the future slots) and gathering.
This patch does this in two passes. One to detect the availability to fold (detect shape and slot change at same offset).
The second pass is used to make sure all Stub information is the same and saves the different shapes (and slots).
The other changes are to support the subsequent patches.

Differential Revision: https://phabricator.services.mozilla.com/D274201

UltraBlame original commit: 066ea7a6368c5a0010e9c35e6727b0d59218995b