Bug 2002625 - wasm: Merge SuspenderObject and SuspenderObjectData. r=yury
Now that we don't require a SuspenderObjectData to be a linked list
element, we can merge it with SuspenderObject. This reduces the number
of registers we need to for operations, and indirections to use.

The JIT routines are largely untouched and can be simplified in a
future patch.

Differential Revision: https://phabricator.services.mozilla.com/D274190

UltraBlame original commit: e5969ca65b8cf0c6c7e8db4c592e4f956016688b