Bug 2002625 - wasm: Use Wasm ABI for builtin thunks. r=yury
Builtin thunks are going to need more scratch registers for the
prologue/epilogue in order to do stack switches. Right now we
are using the sytem ABI for builtin thunks and this leaves us
with only 2 registers for the prologue and 1 register for the
epilogue. It is not feasible to expand this.

We could save/restore non-volatile registers, but this gets
really complicated when the stack is switching to figure out
which stack has the saved register on it.

An easier solution is to move builtin thunks back to using
the wasm ABI, which has no non-volatile registers and therefore
has a good amount of scratch registers that don't alias the
arg/result registers.

We still keep the code to call a builtin using the system ABI
and use that for builtins that don't have a thunk (such as
intgemm). A future commit will expand the number of builtins
which go without a thunk to reduce the overhead of the stack
switch.

A necessary change here is that we must now use saveLive/
restoreLive in some code that is around calling builtins
because we no longer have non-volatile registers.

Differential Revision: https://phabricator.services.mozilla.com/D275544

UltraBlame original commit: fb1b7507031de8551850eff8969e31fbaf29fbf8