Bug 2006816 part 3: Fix TextLeafPoint::GetCaret for LocalAccessible when the caret is deeper than a direct child of the origin container. r=eeejay
Previously, it used HyperTextAccessible::CaretOffset on the origin container, then ToTextLeafPoint on that offset.
However, offsets can only refer to a direct child.
This meant that ToTextLeafPoint was descending into the direct child using offset 0; i.e. returning the start of the child's content.
This was completely incorrect when the caret wasn't at the start of the child's content.
To fix this, first try using the caret cached by SelectionManager, as that avoids redundant transformation of the caret for each descendant.
If that fails, traverse using CaretOffset on each descendant until we can't go any further.

Differential Revision: https://phabricator.services.mozilla.com/D276963

UltraBlame original commit: 41f14f941d6117937cc3a5cd4e21234496dfdf5e