Bug 2006863 - Remove KeyValueStorage.h. r=media-playback-reviewers,chunmin DONTBUILD
It isn't used or built, and uses rkv, that needs to be phased out. It was
referenced in a design document by the team looking at the future of storage as
a user of rkv. I fixed the doc to say they shouldn't consider it a blocker.

Differential Revision: https://phabricator.services.mozilla.com/D276985

UltraBlame original commit: 926019d976e6829c828c8e2b371278c5b6e3e9f3