Bug 2002814 - [devtools] Avoid throwing when seeing a native anonymous node in $$$ console helper. r=devtools-reviewers,ochameau.
The helper was throwing when trying to add an native anonymous node to
the result array as it's seen as a cross origin object.
Wrap the call to Array#push in a try/catch so we avoid throwing
when we're in this case.

Differential Revision: https://phabricator.services.mozilla.com/D277020

UltraBlame original commit: de474b43b073aae6272b4d91508d9bd01d89baef